Decode image samples with single precision calculations only.  This area of the code is performance sensitive and there is no need to convert to and from doubles.  Also, "image_set_gray" is now inline but still not enabled.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11990 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/base/gximage.h
@@ -75,7 +75,7 @@ struct sample_map_s {
      * an 8-bit (unsigned) integer or a frac.
      */
 
-    double decode_factor;
+    float decode_factor;
 
     sample_decoding decoding;
 
@@ -100,7 +100,7 @@ typedef struct sample_map_s sample_map;
   switch ( penum->map[i].decoding )\
   {\
   case sd_none:\
-    cc.paint.values[i] = (sample_value) * (1.0 / 255.0);  /* faster than / */\
+    cc.paint.values[i] = (sample_value) * (1.0f / 255.0f);  /* faster than / */\
     break;\
   case sd_lookup:	/* <= 4 significant bits */\
     cc.paint.values[i] =\
